feat: change blocked routes to chat
