+++ transcendence/srcs_chat/urls.py
@@ -1,31 +1,53 @@
 from django.urls import include, path
 
-from .views import ChatView, GetUpdatedUserListView, ProfileView
+from .views import views
 
 app_name = "srcs_chat"
 
+unblocked_url = [
+    path(
+        "unblocked/",
+        include(
+            [
+                path("", views.UnblockedFormView.as_view(), name="unblocked"),
+                path("<slug:username>", views.UnblockedFormView.as_view()),
+            ]
+        ),
+    )
+]
+
+
+blocked_urls = [
+    path(
+        "blocked/",
+        include(
+            [
+                path("", views.BlockedFormView.as_view(), name="blocked"),
+                path("<slug:username>", views.BlockedFormView.as_view()),
+            ]
+        ),
+    )
+]
+
 urlpatterns = [
-    path("<int:room_id>/", ChatView.as_view(), name="open_chat"),
+    path("<int:room_id>/", views.ChatView.as_view(), name="open_chat"),
     path(
         "create_or_open_chat/<int:user_id>/",
-        ChatView.as_view(),
+        views.ChatView.as_view(),
         name="create_or_open_chat",
     ),
     path(
         "get_updated_user_list/",
-        GetUpdatedUserListView.as_view(),
+        views.GetUpdatedUserListView.as_view(),
         name="get_updated_user_list",
     ),
     path(
         "user/",
         include(
             [
-                # path(
-                #     "blocked/<slug:username>",
-                #     views.BlockedFormView.as_view(),
-                #     name="blocked",
-                # ),
-                path("<slug:username>", ProfileView.as_view(), name="profile"),
+                *blocked_urls,
+                *unblocked_url,
+                path("<slug:username>", views.ProfileView.as_view(), name="profile"),
             ]
         ),
     ),
